rename asset provider and storage provider
