+++ src/providers/export/exportProvider.ts
@@ -1,7 +1,7 @@
 import Guard from "../../common/guard";
 import { IProject } from "../../models/applicationState";
-import { IStorageProvider, StorageProviderFactory } from "../storage/storageProvider";
-import { IAssetProvider, AssetProviderFactory } from "../storage/assetProvider";
+import { IStorageProvider, StorageProviderFactory } from "../storage/storageProviderFactory";
+import { IAssetProvider, AssetProviderFactory } from "../storage/assetProviderFactory";
 
 /**
  * @name - TF Pascal VOC Records Export Asset State
