+++ src/registerProviders.ts
@@ -1,11 +1,11 @@
 import { ExportProviderFactory } from "./providers/export/exportProviderFactory";
 import { TFPascalVOCJsonExportProvider } from "./providers/export/tensorFlowPascalVOC";
 import { VottJsonExportProvider } from "./providers/export/vottJson";
-import { AssetProviderFactory } from "./providers/storage/assetProvider";
+import { AssetProviderFactory } from "./providers/storage/assetProviderFactory";
 import { AzureBlobStorage } from "./providers/storage/azureBlobStorage";
 import { BingImageSearch } from "./providers/storage/bingImageSearch";
 import { LocalFileSystemProxy } from "./providers/storage/localFileSystemProxy";
-import { StorageProviderFactory } from "./providers/storage/storageProvider";
+import { StorageProviderFactory } from "./providers/storage/storageProviderFactory";
 import registerToolbar from "./registerToolbar";
 import { strings } from "./common/strings";
